Add compose schema and fix ajv update dependency
