+++ src/utils/versions/increaseFromLocalVersion.ts
@@ -3,7 +3,11 @@ import {
   readManifest,
   writeManifest
 } from "../../validation/manifest/manifest";
-import { readCompose, writeCompose, updateComposeImageTags } from "../compose";
+import {
+  readCompose,
+  writeCompose,
+  updateComposeImageTags
+} from "../../validation/compose/compose";
 import { checkSemverType } from "../checkSemverType";
 import { ReleaseType } from "../../types";
