+++ src/validation/compose/validateCompose.ts
@@ -1,0 +1,5 @@
+import { Compose } from "../../types";
+
+// Compose schema: https://www.schemastore.org/json/   https://raw.githubusercontent.com/compose-spec/compose-spec/master/schema/compose-spec.json
+
+export async function validateCompose(compose: Compose): void {}
