+++ src/validation/manifest/validateManifest.ts
@@ -1,6 +1,8 @@
 import { CliError, releaseFilesDefaultNames } from "../../params";
 import { Manifest } from "../../types";
-import { validateManifestSchema } from "./validateManifestSchema";
+import Ajv, { ErrorObject } from "ajv";
+import ajvErrors from "ajv-errors";
+import manifestSchema from "./manifest.schema.json";
 
 export function validateManifest(manifest: Manifest): void {
   if (/[A-Z]/.test(manifest.name))
@@ -29,4 +31,90 @@ as ${releaseFilesDefaultNames.avatar} and then remove the 'manifest.avatar' prop
   throw new CliError(
     `Invalid manifest: \n${errors.map(msg => `  - ${msg}`).join("\n")}`
   );
+}
+
+/**
+ * Validates a manifest syncronously. Does NOT throw.
+ * @param manifest
+ * @returns = {
+ *   valid: false|true
+ *   errors: [
+ *     "manifest should have required property 'description'",
+ *     "manifest.image.path should be an non-empty string",
+ *   ]
+ * }
+ */
+function validateManifestSchema(
+  manifest: Manifest
+): { valid: boolean; errors: string[] } {
+  const ajv = new Ajv({ allErrors: true });
+  ajv.errors;
+  ajvErrors(ajv);
+  // Precompile validator
+  const validate = ajv.compile(manifestSchema);
+  const valid = validate(manifest);
+  return {
+    valid: Boolean(valid),
+    errors: validate.errors ? validate.errors.map(processError) : []
+  };
+}
+
+// validate.errors = [
+//   {
+//     keyword: "pattern",
+//     dataPath: "/avatar",
+//     schemaPath: "#/properties/avatar/pattern",
+//     params: { pattern: "^/(ipfs|bzz)/w+$" },
+//     message: 'should match pattern "^/(ipfs|bzz)/w+$"'
+//   },
+//   {
+//     keyword: "pattern",
+//     dataPath: "/image/hash",
+//     schemaPath: "#/properties/image/properties/hash/pattern",
+//     params: { pattern: "^/(ipfs|bzz)/w+$" },
+//     message: 'should match pattern "^/(ipfs|bzz)/w+$"'
+//   },
+//   {
+//     keyword: "type",
+//     dataPath: "/image/size",
+//     schemaPath: "#/properties/image/properties/size/type",
+//     params: { type: "number" },
+//     message: "should be number"
+//   },
+//   {
+//     keyword: "pattern",
+//     dataPath: "/image/size",
+//     schemaPath: "#/properties/image/properties/size/pattern",
+//     params: { pattern: "^d+$" },
+//     message: 'should match pattern "^d+$"'
+//   },
+//   {
+//     keyword: "required",
+//     dataPath: "",
+//     schemaPath: "#/required",
+//     params: { missingProperty: "license" },
+//     message: "should have required property 'license'"
+//   }
+// ];
+
+/**
+ *
+ * @param errorObject from AJV:
+ * {
+ *   keyword: "pattern",
+ *   dataPath: "/avatar",
+ *   schemaPath: "#/properties/avatar/pattern",
+ *   params: { pattern: "^/(ipfs|bzz)/w+$" },
+ *   message: 'should match pattern "^/(ipfs|bzz)/w+$"'
+ * }
+ * @returns errorMessage:
+ * "manifest.avatar should match pattern "^/(ipfs|bzz)/w+$""
+ */
+function processError(
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  errorObject: ErrorObject<string, Record<string, any>, unknown>
+): string {
+  const { schemaPath, message } = errorObject;
+  const path = `manifest${schemaPath}`.replace(new RegExp("/", "g"), ".");
+  return `${path} ${message}`;
 }
